+++ server/common.ts
@@ -4,34 +4,6 @@
  * license information.
  * ===========================================================================*/
 
-// TODO(glen): Move these out to data.
-export const rules = [
-  "ItemLevel", "DropLevel", "Quality", "Rarity", "Class", "BaseType", "Sockets",
-  "LinkedSockets", "SocketGroup", "Height", "Width", "Identified", "Corrupted",
-  "ElderItem", "ShaperItem", "ShapedMap", "SetBorderColor", "SetTextColor",
-  "SetBackgroundColor", "PlayAlertSound", "PlayAlertSoundPositional", "SetFontSize"
-];
-
-// TODO(glen): Move these out to data.
-export const sounds = {
-  numberIdentifier: {
-    min: 1,
-    max: 16
-  },
-  stringIdentifiers: [
-    ["ShGeneral", "General"],
-    ["ShBlessed", "Blessed Orb"],
-    ["ShChaos", "Chaos Orb"],
-    ["ShDivine", "Divine Orb"],
-    ["ShExalted", "Exalted Orb"],
-    ["ShMirror", "Mirror of Kalandra"],
-    ["ShAlchemy", "Orb of Alchemy"],
-    ["ShFusing", "Orb of Fusing"],
-    ["ShRegal", "Regal Orb"],
-    ["ShVaal", "Vaal Orb"]
-  ]
-};
-
 export interface ConfigurationValues {
   baseWhitelist: string[];
   classWhitelist: string[];
@@ -51,4 +23,17 @@ export interface ItemData {
   basesToClasses: { [key: string]: string };
   sortedBases: string[];
   sortedBasesIndices: number[];
+}
+
+export interface FilterData {
+  rules: string[];
+  rarities: string[];
+  booleans: string[];
+  sounds: {
+    numberIdentifier: {
+      min: number,
+      max: number
+    },
+    stringIdentifiers: { [key: string]: string }
+  };
 }