+++ CHANGELOG.md
@@ -2,7 +2,13 @@
 
 The history of changes to the extension.
 
-### Version 1.0.0 (February )
+### Version 0.2.0 (February 13th, 2018)
+
+- Item data added to the project.
+- A basic language server has been added to the project.
+- Autocompletion support added.
+- Whitelist configuration variables added, allowing you to provide additional item classes, item bases, and sound identifiers to the server.
+- Added two other configuration variables, Performance Hints and Always Show Alpha, that will be utilized soon.
 
 ### Version 0.1.0 (February 8th, 2018)